Refactoring: serializers
Moved all serializers from Pyro4.util to Pyro4.serializers.
Added register_serializer to allow adding custom serializers.
Refactored dict_to_class to allow using custom serializers.

The classname of serializers is changed and is transmitted over the
wire so Pyro4 with this patch would not be able to communicate with an
older version.
(eg. Pyro4.utils.SerpentSerializer -> Pyro4.serializers.SerpentSerializer)
On the receive side, this can be handled by adding support for the old classname,
but on the send side, the only way would be to handle differently based on protocol version.